Table cleanup before subRows PR
