+++ src/core/Table/actionHandlers/index.ts
@@ -1,0 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+export { onExpandHandler } from './expandHandler';
+export { onFilterHandler } from './filterHandler';
+export { onSelectHandler } from './selectHandler';
