+++ src/core/Table/hooks/index.ts
@@ -1,0 +1,6 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+export { EXPANDER_CELL_ID, useExpanderCell } from './useExpanderCell';
+export { SELECTION_CELL_ID, useSelectionCell } from './useSelectionCell';
